plat: imx95: Assign all the GPIO pins to non-secure world by default
For the Variscite DART-MX95 we are building system manager image based on the imx95alt configuration.
This is necessary to give the permission to the Cortex-M7 core.

When using the system manager configuration based on mx95alt it seems that LM1 which contain Cortex A55
run in secure mode so we need to give secure access to GPIOs to use them from Cortex A55.

To set the GPIOs pins/pads for secure access the PCNS (offset 0x10) and ICNS (offset 0x14) registers
have been changed accordingly.

For further details see the RM par. 40.7.1.5 (Pin Control Nonsecure (PCNS)) and 40.7.1.6 (Interrupt Control Nonsecure (ICNS))

Signed-off-by: FrancescoFerraro <francesco.f@variscite.com>